+++ packages/types/src/api/CanvasInfoRequest.d.ts
@@ -1,4 +1,4 @@
-import type { CanvasInfo } from "..";
+import type { CanvasInfo } from "@/canvasInfo";
 
 export interface Params {
   canvasId: CanvasInfo["id"];
