+++ packages/types/src/api/EventRequest.d.ts
@@ -1,4 +1,4 @@
-import type { BlurpleEvent } from "..";
+import type { BlurpleEvent } from "@/event";
 
 export interface Params {
   canvasId: BlurpleEvent["id"];
