+++ packages/types/src/api/HistoryRequest.d.ts
@@ -1,4 +1,4 @@
-import type { PixelHistory } from "..";
+import type { PixelHistory } from "@/pixelHistory";
 
 export interface Params {
   canvasId: number;
